Knockback, Player Spawning, Scene work.
Knockback sort of working. Still buggy, need to identify what causes inconsistencies in knockback and fix them. Player spawning has been reworked to spawn players in a corner. If the corner is full , pick a different one. The scene has been changed for a more competitive arena with more branches to grapple.